+++ e2e/nx-boot-maven-e2e/tests/nx-boot-maven.spec.ts
@@ -1470,4 +1470,148 @@ describe('nx-boot-maven e2e', () => {
       expect(err).toBeFalsy();
     }
   }, 1200000);
+
+  it('should create an application with a simple name', async () => {
+    const appName = uniq('boot-maven-app-');
+    const appDir = 'deep/subdir';
+
+    await runNxCommandAsync(
+      `generate @jnxplus/nx-boot-maven:application ${appName} --simpleName --tags e2etag,e2ePackage --directory ${appDir} --groupId com.jnxplus --projectVersion 1.2.3 --packaging war --configFormat .yml`
+    );
+
+    expect(() =>
+      checkFilesExist(
+        `apps/${appDir}/${appName}/pom.xml`,
+        `apps/${appDir}/${appName}/src/main/resources/application.yml`,
+        `apps/${appDir}/${appName}/src/main/java/com/jnxplus/deep/subdir/${names(
+          appName
+        ).className.toLocaleLowerCase()}/${
+          names(appName).className
+        }Application.java`,
+        `apps/${appDir}/${appName}/src/main/java/com/jnxplus/deep/subdir/${names(
+          appName
+        ).className.toLocaleLowerCase()}/HelloController.java`,
+
+        `apps/${appDir}/${appName}/src/test/resources/application.yml`,
+        `apps/${appDir}/${appName}/src/test/java/com/jnxplus/deep/subdir/${names(
+          appName
+        ).className.toLocaleLowerCase()}/HelloControllerTests.java`
+      )
+    ).not.toThrow();
+
+    // Making sure the pom.xml file contains the good informations
+    const pomXml = readFile(`apps/${appDir}/${appName}/pom.xml`);
+    expect(pomXml.includes('com.jnxplus')).toBeTruthy();
+    expect(pomXml.includes('1.2.3')).toBeTruthy();
+    expect(pomXml.includes('war')).toBeTruthy();
+    expect(pomXml.includes('spring-boot-starter-tomcat')).toBeTruthy();
+
+    //should add tags to project.json
+    const projectJson = readJson(`apps/${appDir}/${appName}/project.json`);
+    expect(projectJson.tags).toEqual(['e2etag', 'e2ePackage']);
+
+    const buildResult = await runNxCommandAsync(
+      `build ${appName} --mvnArgs='--no-transfer-progress'`
+    );
+    expect(buildResult.stdout).toContain('Executor ran for Build');
+
+    const testResult = await runNxCommandAsync(`test ${appName}`);
+    expect(testResult.stdout).toContain('Executor ran for Test');
+
+    const lintResult = await runNxCommandAsync(`lint ${appName}`);
+    expect(lintResult.stdout).toContain('Executor ran for Lint');
+
+    const formatResult = await runNxCommandAsync(
+      `format:check --projects ${appName}`
+    );
+    expect(formatResult.stdout).toContain('');
+
+    //graph
+    const depGraphResult = await runNxCommandAsync(
+      `dep-graph --file=dep-graph.json`
+    );
+    expect(depGraphResult.stderr).not.toContain(
+      'Failed to process the project graph'
+    );
+    const depGraphJson = readJson('dep-graph.json');
+    expect(depGraphJson.graph.dependencies[appName]).toContainEqual({
+      type: 'static',
+      source: appName,
+      target: parentProjectName,
+    });
+
+    const process = await runNxCommandUntil(
+      `serve ${appName} --args="-Dspring-boot.run.profiles=test"`,
+      (output) => output.includes(`Tomcat started on port(s): 8080`)
+    );
+
+    // port and process cleanup
+    try {
+      await promisifiedTreeKill(process.pid, 'SIGKILL');
+      await killPorts(8080);
+    } catch (err) {
+      expect(err).toBeFalsy();
+    }
+  }, 1200000);
+
+  it('should create a library with a simple name', async () => {
+    const libName = uniq('boot-maven-lib-');
+    const libDir = 'deep/subdir';
+
+    await runNxCommandAsync(
+      `generate @jnxplus/nx-boot-maven:library ${libName} --simpleName --directory ${libDir} --tags e2etag,e2ePackage --groupId com.jnxplus --projectVersion 1.2.3`
+    );
+
+    expect(() =>
+      checkFilesExist(
+        `libs/${libDir}/${libName}/pom.xml`,
+        `libs/${libDir}/${libName}/src/main/java/com/jnxplus/deep/subdir/${names(
+          libName
+        ).className.toLocaleLowerCase()}/HelloService.java`,
+        `libs/${libDir}/${libName}/src/test/java/com/jnxplus/deep/subdir/${names(
+          libName
+        ).className.toLocaleLowerCase()}/TestConfiguration.java`,
+        `libs/${libDir}/${libName}/src/test/java/com/jnxplus/deep/subdir/${names(
+          libName
+        ).className.toLocaleLowerCase()}/HelloServiceTests.java`
+      )
+    ).not.toThrow();
+
+    // Making sure the pom.xml file contains the good information
+    const pomXml = readFile(`libs/${libDir}/${libName}/pom.xml`);
+    expect(pomXml.includes('com.jnxplus')).toBeTruthy();
+    expect(pomXml.includes('1.2.3')).toBeTruthy();
+
+    //should add tags to project.json
+    const projectJson = readJson(`libs/${libDir}/${libName}/project.json`);
+    expect(projectJson.tags).toEqual(['e2etag', 'e2ePackage']);
+
+    const buildResult = await runNxCommandAsync(`build ${libName}`);
+    expect(buildResult.stdout).toContain('Executor ran for Build');
+
+    const testResult = await runNxCommandAsync(`test ${libName}`);
+    expect(testResult.stdout).toContain('Executor ran for Test');
+
+    const lintResult = await runNxCommandAsync(`lint ${libName}`);
+    expect(lintResult.stdout).toContain('Executor ran for Lint');
+
+    const formatResult = await runNxCommandAsync(
+      `format:check --projects ${libName}`
+    );
+    expect(formatResult.stdout).toContain('');
+
+    //graph
+    const depGraphResult = await runNxCommandAsync(
+      `dep-graph --file=dep-graph.json`
+    );
+    expect(depGraphResult.stderr).not.toContain(
+      'Failed to process the project graph'
+    );
+    const depGraphJson = readJson('dep-graph.json');
+    expect(depGraphJson.graph.dependencies[libName]).toContainEqual({
+      type: 'static',
+      source: libName,
+      target: parentProjectName,
+    });
+  }, 1200000);
 });