+++ packages/nx-boot-gradle/src/generators/library/schema.d.ts
@@ -4,6 +4,7 @@ export interface NxBootGradleLibGeneratorSchema {
   name: string;
   tags?: string;
   directory?: string;
+  simpleName?: boolean;
   language: LanguageType;
   groupId: string;
   packageNameType: PackageNameType;