+++ packages/nx-quarkus-maven/src/generators/application/schema.d.ts
@@ -4,6 +4,7 @@ export interface NxQuarkusMavenAppGeneratorSchema {
   name: string;
   tags?: string;
   directory?: string;
+  simpleName?: boolean;
   language: LanguageType;
   groupId: string;
   packageNameType: PackageNameType;